story-app-fastapi-backend/requirements.txt
fastapi==0.95.2
pydantic==1.10.7
SQLAlchemy==2.0.15
uvicorn==0.22.0
python-dotenv==0.21.1
openai==0.27.8